Add SugarUtils::File.read/write/flock_exclusive/flock_shared
The original goal was to add the read and write methods. These will
properly flock and timeout, just as the JSON read and write methods do.

In the process I have also separated the flock_exclusive and
flock_shared, in order to hide the specific details of how these are
selected.