Fix error handling on SugarUtils::File.read_json
When read_json is passed a 'raise_on_missing:false' option it was
supposed to return {}. Unfortunately, I did not handle this scenario
properly which would result in the MultiJson.load being sent a {}
value.

Which because it was not a String, as expected, would raise a confusing
error inside of which every JSON engine MultiJson is proxying for.

.read_json has been updated to check the result of .read before trying
to parse it, and I have added a spec example to check this scenario.