Add Sugar::Utils::File.append
For a user, this will help their code clearly indicate whether they are
creating a brand new file (.write) or should append to an existing file
(.append)

Please note: that the :mode and :perm options are alias for each other,
allowing the user to specified what permissions should be assinged to
the file. The other methods will be updated to follow these scheme in
the future.

On the development side, this will make it easier to handle scenarios
which are specific to either appending or writing.

Closes #6