Disable RSpec/EmptyLineAfterExample
I think that it is OK to put some examples side-by-side.
I especially want to be able to but an it and its_side_effects_are
together. There are several examples of this in the code, and it looks
like this:

```
it { expect_not_to_raise_error }
its_side_effects_are do
  expect(...).to be_something
end
```