Update keyword argument related code
There are 2 places in the code which will fail in on 3.x
because of the implicit Hash at the end of a method call is
deprecated in favour of keyword arguments.

Because keyword arguments were introduced all the way back in Ruby 2.0,
we can just switch and the code will work with all the versions that we
want to support. :tada:

This includes:
* use the `**` operator to pass keyword arguments to FileUtils.touch
* add explicit Hash brackets to an allow/receive in the specs, so that
  Ruby will continue to expect the options Hash instead of keyword
  arguments