Add Code Climate coverage reporting to CI
Using the `paambatti/codeclimate-action` which will automatically download
and execute the Code Climate test reporter.

* https://github.com/paambaati/codeclimate-action
* https://github.com/marketplace/actions/code-climate-coverage-action
* https://github.com/codeclimate/test-reporter

Additionally, I found that the test reporter does not reliably read the
simplecov JSON format, even though it will attempt to use it by default.

So I have added the lcov formatter, and configured the reporter to use
it.

* https://github.com/fortissimo1997/simplecov-lcov

Finally, the CC_TEST_REPORTER_ID is stored as a repository secret
(i.e., `${{ secrets.CC_TEST_REPORTER_ID }}`) and will need to be added
before this action will work as expected.