Upgrade minimum required Ruby to v3.1.2
I am considering this the correct minimum Ruby because it is the Ruby
which is available in Debian 12 (bookworm), which is current in January
2025.

* https://www.debian.org/releases/
* https://packages.debian.org/bookworm/ruby/

Some rubocop refactoring, such as:
* remove the bundler dependency, becuase it is built in now

Github Action script and CI updates:
* drop the unsupported Rubies and added 3.4
* make some updates related to testing Ruby v3.4
  - ignore white space when comparing strignfied Hash values in
    Cucumber because white space is added to this output in 3.5
    * https://rubyreferences.github.io/rubychanges/3.4.html#inspect-rendering-have-been-changed
    * https://bugs.ruby-lang.org/issues/20433
* make some updates related to testing Ruby head (i.e. v3.5)
  - add `logger` as a development dependency because it is dropped from
    the standard library in Ruby v3.5
  - add `irb` as a development dependency becuase it is needed by
    `faker` in Ruby 3.5
* remove ruby version specification from the `coverage` step, because it
  should use the default/minimum Ruby specified in the `.ruby-version`
* update action/checkout step to v4
* update paambatti/codeclimate to v9.0.0

Update some Cucumber steps to handle the addition of more whitespace to
stringified Hashes in Ruby 3.4. Any whitespace will be ignored on this
check for any version.

Try to exclude running the Cucumber/Aruba feature integration tests on
Ruby head. Neither Cucumber or Aruba properly support Ruby head/3.5 at the
moment.